+++ lambda/readings/readings/utils/sqlalchemy.py
@@ -2,13 +2,13 @@ import os
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-
 ENGINE = create_engine('mysql+pymysql://{username}:{password}@{host}/{database}'.format(
     username=os.getenv('MYSQL_USERNAME'),
     password=os.getenv('MYSQL_PASSWORD'),
     host=os.getenv('MYSQL_HOST'),
     database=os.getenv('MYSQL_DATABASE'),
 ))
+
 SESSION = sessionmaker(bind=ENGINE)
